requirements.txt
flask
flask-cors
flask-sqlalchemy
requests
google-generativeai
python-dotenv
gunicorn
